added two new options to PasDoc:
--link-gv-uses <extension>
--link-gv-classes <extension>

They work only in html output (It seems as if there was already something similar in LaTeX output, but I haven't checked really).

The simply add a link to the navigation frame to the "Unit dependency graph" and "Classes hierarchy graph" respectively. <extension> is the file extension to use. It is up to the user to generate these files using the 'dot' program from GraphViz.

I have also added a (unix-) shell script to call pasdoc and dot to generate PasDoc documentation on the sources.

The English and German translations for the link captions have been added, we need the other languages.

Bugfixes:
When generating the GraphViz files in WriteGVUses and
WriteGVClasses there were four more instances where WriteConverted was used instead of WriteDirect.

Other changes:
Writing the index.html and navigation.html files has been moved from the WriteBinaryFiles method to a new WriteFramesetFiles. Also a subprocedure has been added for writing the links to the navigation frame using an array constant to determine the captions rather than having a case statement. (I hope this is supported by all compilers, I am using Kylix 3 and I know that it works with Delphi 3 up).


git-svn-id: https://svn.code.sf.net/p/pasdoc/code/trunk@239 e0417a13-a44b-4b77-80e5-317e04f1597d