Change Ancestors to TStringPairVector.
We will need 2nd string to keep the specialization "<foo,bar>" string which may be given in ancestors list for each ancestor. But it's not yet parsed there --- the 2nd string is always empty now.

Small improvements to TStringPairVector were required for this.

Cache version bumped, as TStringPairVector is stored differently.

BTW, missing svn:eol-style for PasDoc_GenSimpleXML.pas.


git-svn-id: https://svn.code.sf.net/p/pasdoc/code/trunk@1302 e0417a13-a44b-4b77-80e5-317e04f1597d