Fix compilation with FPC 3.0.4
FPC 3.0.4 (unlike 3.3.1) doesn't have SplitString in StrUtils. But
using {$IF NOT DECLARED(SplitString)} in both interface and
implementation doesn't work as expected, as it is not true the 2nd
time (in implementation) so compilation was failing with

  PasDoc_Types.pas(87,10) Error: Forward declaration not solved "SplitString(const AnsiString;const AnsiString):{Dynamic} Array Of AnsiString;"

Also, need to use PosEx in FPC. There is no overloaded Pos with StartIndex.

Also, TStringArray can be an alias to TStringDynArray in FPC Types
unit.

Also, "t: string;" was unused.

See https://github.com/pasdoc/pasdoc/pull/45/