Playing around, but left commented out, logic to "On Unix accept also relative paths with backslash (Windows)"
This is not really correct (on Unix backslash is a valid filename
character). Although it seems useful (it is probably more common case
that someone tries to use backslash as directory separator with
PasDoc, than someone having actual backslash in a directory
name). Maybe it should be toggled with a command-line switch.

Inspired by testing https://github.com/pasdoc/pasdoc/issues/122 on Linux.